+++ bson/_cbsonmodule.h
@@ -73,6 +73,11 @@ typedef struct codec_options_t {
     PyObject* options_obj;
     unsigned char is_raw_bson;
     unsigned char is_dict_class;
+    /* Decode-buffer state for zero-copy RawBSONDocument slices */
+    PyObject* buffer_owner; /* borrowed */
+    const char* view_base;
+    Py_ssize_t view_len;
+    PyObject* top_view; /* owned */
 } codec_options_t;
 
 /* C API functions */